Heretic: add support for ZDBSP, DeePBSP and Hexen nodes

+++ src/heretic/p_local.h
@@ -248,8 +248,8 @@ void P_RadiusAttack(mobj_t * spot, mobj_t * source, int damage);
 // ***** P_SETUP *****
 
 extern byte *rejectmatrix;      // for fast sight rejection
-extern long *blockmaplump;     // offsets in blockmap are from here
-extern long *blockmap;
+extern int32_t *blockmaplump;     // offsets in blockmap are from here
+extern int32_t *blockmap;
 extern int bmapwidth, bmapheight;       // in mapblocks
 extern fixed_t bmaporgx, bmaporgy;      // origin of block map
 extern mobj_t **blocklinks;     // for thing chains
